Fix hard coded size for various vectors
There were several instances of hardcoding size data. This commit uses programmatic
means to get the correct size, to ensure issues do not occur in the future.

Spiffed up where the imgui windows are.
Moved the RNG to be a class variable, so its not being constructed every frame
Moved scene init to before the rvpt init function